==> NLP project/practicing/1. essay_url_crawling.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.service import Service
import time

# 드라이버 경로 설정
driver_path = r'.\chromedriver-win64\chromedriver.exe'

options = webdriver.ChromeOptions()
options.add_argument('--headless')
options.add_argument('--no-sandbox')
options.add_argument('--disable-dev-shm-usage')
s = Service(driver_path)
driver = webdriver.Chrome(service=s, options=options)

# 합격 자소서 게시판 -> 지원분야 : 개발·데이터 -> 전문가 분석 보기 -> 평점 높은순 : 총 323건 
url_list = ['https://www.jobkorea.co.kr/starter/PassAssay?FavorCo_Stat=0&Pass_An_Stat=1&OrderBy=2&EduType=0&WorkType=0&schPart=10031&isSaved=1&Page=' + str(x) for x in range(1, 18)]
links = []
for url in url_list:
    driver.get(url)
    time.sleep(3)
    question_items = driver.find_elements(By.CSS_SELECTOR, 'div.item.question')
    for item in question_items:
        a_tag = item.find_element(By.TAG_NAME, 'a')
        href = a_tag.get_attribute('href')
        links.append(href)

driver.quit()

len(links) # 자소서가 323개 모두 가져와졌는지 확인

# url list를 csv 파일로 저장
import csv

file_name = 'urls.csv'
with open(file_name, 'a', newline='') as file:
    writer = csv.writer(file)
    for link in links:
        writer.writerow([link])
        
print(f'{file_name} 작성 완료')
