==> NLP project/practicing/2. url_kr_cv_3.ipynb <==
from selenium import webdriver
import bs4
import csv

# 크롬 드라이버 설정
driver = webdriver.Chrome()

# 자소서 url 리스트
file_path = 'urls.csv'
with open(file_path, "r", encoding="utf-8-sig") as f:
    url_reader = csv.reader(f)
    urls = [row[0] for row in url_reader][160:240]


def crawling_url(url):
    driver.get(url)
    response = driver.page_source
    one_page = bs4.BeautifulSoup(response, 'html.parser')
    
    # 불필요 요소 제거
    for p_element in one_page.find_all('p', {'class': 'txSpllChk'}):
        p_element.extract()

    for span_element in one_page.find_all('span', {'class': 'sup'}):
        span_element.extract()

    for a_element in one_page.find_all('a'):
        a_element.extract()
        
    # 공백 제거 및 텍스트만 추출
    kwd_list = []
    for text in one_page.find_all('div', {'class': 'tx'}):
        kwd_list.append(text.get_text(separator=' ', strip=True))
    
    return kwd_list

all_kwd_lists = []
for kwd_url in urls:
    kwd_list = crawling_url(kwd_url)
    all_kwd_lists.append(kwd_list)

driver.quit()

# 크롤링한 데이터를 텍스트 파일로 저장 후 내용 확인
output_file_path = 'crawled_data.txt'

with open(output_file_path, 'w', encoding='utf-8') as file:
    writer = csv.writer(file)
    for i, kwd_list in enumerate(all_kwd_lists):
        writer.writerow(kwd_list)

# 크롤링한 데이터를 CSV 파일로 저장
output_file_path = 'crawled_data0.csv'

with open(output_file_path, 'w', newline='', encoding='utf-8') as file:
    writer = csv.writer(file)
    
    for i, kwd_list in enumerate(all_kwd_lists):
        writer.writerow(kwd_list)


# join
join_list = []
input_file_path = 'crawled_data0.csv'
with open(input_file_path, 'r', encoding='utf-8') as file:
    reader = csv.reader(file)
    
    for row in reader:
        sen = ' '.join(row)
        join_list.append(sen)

# 새 파일로 저장 
output_file_path = 'crawled_data1.csv'
with open(output_file_path, 'w', encoding='utf-8', newline='') as file:
    writer = csv.writer(file)
    
    for sen in join_list:
        writer.writerow([sen])
        
print(f"{output_file_path} 저장 완료")

file_path = r'C:/Users/user/Desktop/hansol/3rd_project/NLP_project/crawled_data0.csv'

with open(file_path, 'r', encoding='utf-8') as f :
 file_read = csv.reader(f)
 
 cnt = 0
 for i, row in enumerate(file_read) :
    val=[len(value) for value in row]
    print(f'{i+1}행 : {sum(val)}')
    
    if sum(val) > 2990 :
        cnt +=1

print(f'\n2990자가 넘는 자소서 숫자 : {cnt}') # 80개 중 23개의 자소서가 3천자를 넘음
