==> NLP project/practicing/5. over 6k check.ipynb <==
import csv

file_path = r'C:\\Users\\user\\Desktop\\hansol\\3rd_project\\NLP_new_\\NLP-project\\NLP project\\crawling_data\\(new)korean1-317.csv'

n = 6000

with open(file_path, 'r', encoding='utf-8') as f :
 file_read = csv.reader(f)
 
 cnt = 0
 for i, row in enumerate(file_read) :
    val=[len(value) for value in row]
    print(f'{i+1}행 : {sum(val)}')
    
    if sum(val) > n :
        cnt +=1

print(f'\n{n}자가 넘는 자소서 숫자 : {cnt}')

n = 6000

rows_exceeding_n = []

with open(file_path, 'r', encoding='utf-8') as f:
    file_read = csv.reader(f)
    
    for i, row in enumerate(file_read):
        val = [len(value) for value in row]
        total_length = sum(val)
        
        if i == 51 or i == 61 :
            print(f'{i+1}행 : {total_length}')
        
        if total_length > n:
            rows_exceeding_n.append(i + 1)  # 행 번호는 1부터 시작하므로 +1

print(f'\n{n}자가 넘는 자소서 숫자 : {len(rows_exceeding_n)}')
print(f'{n}자가 넘는 자소서의 행 번호: {rows_exceeding_n}')
